+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## [0.0.1] - 25/08/2019.
 
 * First release: Added application configuration.
+
+## [0.0.1] - 26/08/2019.
+
+* Changed localizationsDelegate field name to localizationsDelegates.
+
+* Changed type of property from single delegate to list of localization delegates.
+This allow an application configuration to provide multiple localization delegates
+